ONL-6717:refactor:use symbols, discards enums and move metro item inside metroline
