--- src/enums/metroline-item-status.js
@@ -1,3 +1,0 @@
-export const NEXT = 'next';
-export const ACTIVE = 'active';
-export const COMPLETED = 'completed';
