+++ src/symbols/provide-inject.js
@@ -1,0 +1,1 @@
+export const METROLINE_SYMBOL = Symbol('metroline');
